comments codestyled, no-content-type test
