+++ requirements.txt
@@ -3,4 +3,3 @@ dj-database-url==0.3.0
 django-jsonfield==0.9.13
 git+https://github.com/sarumont/py-trello.git@766c90dc1dacd2e3fcfc579079a6aab38be43aef
 psycopg2==2.5.4
-mock==1.0.1
