+++ test_app/templates/trello_webhooks/addAttachmentToCard.html
@@ -1,0 +1,3 @@
+{% load test_app_tags %}
+<strong>{{action.memberCreator.initials}}</strong> added attachment <strong><a href="{{action.data.attachment.url}}">{{action.data.attachment|render_attachment}}</a></strong>
+{% include 'trello_webhooks/partials/card_link.html' %}
