+++ test_app/templatetags/__init__.py
@@ -1,0 +1,4 @@
+# test_app.templatetags package
+# do this to force coverage to include template_tags
+import test_app_tags
+__all__ = ['test_app_tags']
