+++ tox.ini
@@ -2,7 +2,7 @@
 envlist = django171_py27
 
 [testenv]
-commands=python manage.py test trello_webhooks
+commands=python manage.py test trello_webhooks test_app
 setenv =
     DJANGO_SETTINGS_MODULE=test_app.test_settings
     PYTHONPATH={toxinidir}
@@ -15,6 +15,7 @@ setenv =
 basepython = python2.7
 deps =
     mock==1.0.1
+    httpretty==0.8.10
     coverage==3.7.1
     django-nose==1.2
     -rrequirements.txt
