+++ trello_webhooks/templates/trello_webhooks/addAttachmentToCard.html
@@ -1,3 +1,2 @@
-{% load attachments %}
-<strong>{{action.memberCreator.initials}}</strong> added attachment <strong><a href="{{action.data.attachment.url}}">{{action.data.attachment|render_attachment}}</a></strong>
+<strong>{{action.memberCreator.initials}}</strong> added attachment "<strong><a href="{{action.data.attachment.url}}">{{action.data.attachment.name}}</a></strong>"
 {% include 'trello_webhooks/partials/card_link.html' %}
